requirements.txt
kafka-python
google-cloud-storage
flask

